% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddClinicFromFile.R
\name{AddClinicFromFile}
\alias{AddClinicFromFile}
\title{AddClinicFromFile to Meta object}
\usage{
AddClinicFromFile(
  Metadata,
  ClinicFile = NULL,
  setSamplesID.Column = NULL,
  setPatientID.Column = NULL,
  setCellsBarcode.Column = NULL,
  name = NULL,
  type = c("Samples", "Patients", "Cells"),
  force.replace = F,
  Export = F,
  mergeToClinic = NULL,
  mergeBy = NULL,
  join = c("left_join", "full_join")
)
}
\arguments{
\item{Metadata}{Meta object}

\item{ClinicFile}{name file of interest in path directory, could be multiple names. c("a.csv","b.csv")}

\item{setSamplesID.Column}{a character string : column to fetch SamplesID from Count matrix.}

\item{setPatientID.Column}{a character string : column to fetch PatientsID from Count matrix.}

\item{setCellsBarcode.Column}{a character string : column to fetch CellBarcode from Count matrix.}

\item{name}{if loca=True, names to apply in Metadata object slot}

\item{type}{c("Samples", "Patients") To specify if clinical data refers to samples or patients data. Importat if mutliple samples for one patient.}

\item{force.replace}{set as F. T : replace an already object with the same name}

\item{Export}{TRUE or FALSE. If data to be Exported, set T.}

\item{mergeToClinic}{default NULL. name of loaded clinical data in Metadata. Merge the newly added clinical data to a already loaded data clinic with existing clincial data : full_join by rownames.}

\item{mergeBy}{colname using for merging clinical data.}

\item{join}{c("left_join", "full_join")}
}
\value{
a data.frame in the Meta Object
}
\description{
AddClinicFromFile to Meta object
}
\examples{
"none"
}
