% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddExpressionMatrixMArray.R
\name{AddExpressionMatrixMArray}
\alias{AddExpressionMatrixMArray}
\title{AddExpressionMatrix MicroArray to Meta object}
\usage{
AddExpressionMatrixMArray(
  Metadata = NULL,
  name = NULL,
  ExpressionMatrix = NULL,
  force.replace = F
)
}
\arguments{
\item{Metadata}{Meta object}

\item{name}{if loca=True, names to apply in Metadata object slot}

\item{ExpressionMatrix}{name file of interest in path directory}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
a data.frame in the Meta Object
}
\description{
AddExpressionMatrix MicroArray to Meta object
}
\examples{
"none"
}
