% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddExpressionMatrixRNAseq.R
\name{AddExpressionMatrixRNAseq}
\alias{AddExpressionMatrixRNAseq}
\title{AddExpressionMatrixRNAseq to Meta object}
\usage{
AddExpressionMatrixRNAseq(
  Metadata = NULL,
  name = NULL,
  ExpressionMatrix = NULL,
  force.replace = F
)
}
\arguments{
\item{Metadata}{Meta object}

\item{name}{if loca=True, names to apply in Metadata object slot}

\item{ExpressionMatrix}{an object or a character string}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
a data.frame in the Meta Object
}
\description{
AddExpressionMatrixRNAseq to Meta object
}
\examples{
"none"
}
