% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddExpressionMatrixSC.R
\name{AddExpressionMatrixSC}
\alias{AddExpressionMatrixSC}
\title{AddExpressionMatrixSC from .mtx to Meta object}
\usage{
AddExpressionMatrixSC(
  Metadata = NULL,
  Cell.file = NULL,
  Genes.file = NULL,
  name,
  ExpressionMatrix = NULL,
  force.replace = F
)
}
\arguments{
\item{Metadata}{Metaobject}

\item{Cell.file}{an object or a character string name of single cell cell annotation file specific to expression matrix. Not mandatory.}

\item{Genes.file}{an object or a character string name of single cell gene annotation file specific to expression matrix. Not mandatory.}

\item{name}{if loca=True, names to apply in Metadata object slot}

\item{ExpressionMatrix}{an object or a character string name file of interest in path directory from .mtx exclusive}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
a data.frame in the Meta Object
}
\description{
AddExpressionMatrixSC from .mtx to Meta object
}
\examples{
"none"
}
