% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddExpressionMatrixTCGA.R
\name{AddExpressionMatrixTCGA}
\alias{AddExpressionMatrixTCGA}
\title{AddExpressionMatrixTCGA from TCGA to Meta object}
\usage{
AddExpressionMatrixTCGA(
  Metadata = NULL,
  query,
  data.norm,
  name,
  name.local.file = NULL,
  force.replace = F
)
}
\arguments{
\item{Metadata}{Meta object}

\item{query}{query file from TCGAimportExpression function}

\item{data.norm}{a character "Raw" ,"TPM", "FKPM"}

\item{name}{if loca=True, names to apply in Metadata object slot}

\item{name.local.file}{name file of interest in path directory}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
a data.frame in the Meta Object
}
\description{
AddExpressionMatrixTCGA from TCGA to Meta object
}
\examples{
"none"
}
