% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddgeneAnnot.R
\name{AddgeneAnnot}
\alias{AddgeneAnnot}
\title{AddgeneAnnot to Meta}
\usage{
AddgeneAnnot(Meta, gtf.file.dir, gtf.files)
}
\arguments{
\item{Meta}{Meta object to add geneAnnotation}

\item{gtf.file.dir}{set xorking directory to path to gtf.file}

\item{gtf.files}{gtf file to load}
}
\value{
return a Meta list of dataframe with geneAnntotation
}
\description{
AddgeneAnnot to Meta
}
\examples{
"none"

}
