% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddgeneAnnotFromObject.R
\name{AddgeneAnnotFromObject}
\alias{AddgeneAnnotFromObject}
\title{AddgeneAnnotFromObject to Meta}
\usage{
AddgeneAnnotFromObject(Meta, object, force.replace = F)
}
\arguments{
\item{Meta}{Meta object to add geneAnnotation}

\item{object}{object to add}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
return a Meta list of dataframe with geneAnntotation
}
\description{
AddgeneAnnotFromObject to Meta
}
\examples{
"none"

}
