% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddgeneAnnotGtf.R
\name{AddgeneAnnotGtf}
\alias{AddgeneAnnotGtf}
\title{AddgeneAnnotGtf to Metadata}
\usage{
AddgeneAnnotGtf(Metadata, gtf.files, force.replace = F)
}
\arguments{
\item{Metadata}{Metadata object to add geneAnnotation}

\item{gtf.files}{gtf file to load}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
return a Metadata list of dataframe with geneAnntotation
}
\description{
AddgeneAnnotGtf to Metadata
}
\examples{
"none"

}
