% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddKeyLexic.R
\name{AddKeyLexic}
\alias{AddKeyLexic}
\title{AddKeyLexic}
\usage{
AddKeyLexic(lexic = NULL, key = NULL, value = NULL)
}
\arguments{
\item{lexic}{PatientLexic or SamplesLexic}

\item{key}{a character string: name to add into lexic}

\item{value}{a character string: target column in "raw" clinical data to rename as 'key' in cleaned clinical data.}
}
\value{
a Lexic
}
\description{
AddKeyLexic
}
\examples{
"None"
}
