% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddObjetToMeta.R
\name{AddObjetToMeta}
\alias{AddObjetToMeta}
\title{AddObjetToMeta}
\usage{
AddObjetToMeta(
  Meta,
  object,
  name,
  Data.type = c("Expression.Matrix", "Patient.Clinical.data", "Samples.Clinical.data",
    "geneAnnotation.file"),
  Raw = c("Yes", "No")
)
}
\arguments{
\item{Meta}{a Meta Object to fill}

\item{object}{an object to add}

\item{name}{name to apply in Meta objet list}

\item{Data.type}{attribute c("Expression.Matrix","Patient.Clinical.data", "Samples.Clinical.data","geneAnnotation.file" )}

\item{Raw}{attribute c("Yes", "No")}
}
\value{
a Meta Object
}
\description{
AddObjetToMeta
}
