% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddKeyLexic.R
\name{addSeveralKeysToLexic}
\alias{addSeveralKeysToLexic}
\title{addSeveralKeysToLexic}
\usage{
addSeveralKeysToLexic(vector = NULL, lexic = NULL)
}
\arguments{
\item{vector}{a list of c(key=value) for}

\item{lexic}{lexic to add}
}
\value{
a lexic
}
\description{
addSeveralKeysToLexic
}
\examples{
"not"
}
