% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charac.to.Num.R
\name{Charac.to.Num}
\alias{Charac.to.Num}
\title{Charac.to.Num  transform only numbers column from character to numeric}
\usage{
Charac.to.Num(clinic)
}
\arguments{
\item{clinic}{clinic data to perform}
}
\description{
Charac.to.Num  transform only numbers column from character to numeric
}
