% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckCleaning.R
\name{CheckCleaning}
\alias{CheckCleaning}
\title{CheckCleaning function for checking numerical values in Clean clinical data, as all numerical.}
\usage{
CheckCleaning(Metadata)
}
\arguments{
\item{Metadata}{.}
}
\description{
CheckCleaning function for checking numerical values in Clean clinical data, as all numerical.
}
\examples{
"non"
}
