% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckLexicDups.R
\name{CheckLexicDups}
\alias{CheckLexicDups}
\title{CheckLexicDups}
\usage{
CheckLexicDups(Lexic)
}
\arguments{
\item{Lexic}{Lexic to check if duplicated values (Targeted column) are in the lexic items.}
}
\value{
nothing
}
\description{
CheckLexicDups
}
\examples{
"non"
}
