% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckLocalDatabase.R
\name{CheckLocalDatabase}
\alias{CheckLocalDatabase}
\title{CheckLocalDatabase}
\usage{
CheckLocalDatabase(Local.Database.Path, Databasename = "DataBaseSummary.txt")
}
\arguments{
\item{Databasename}{name of the database.txt summary file}

\item{Local.Data.base.Path}{path to parent directory of the database.}
}
\value{
a .text tab delimited database summary
}
\description{
CheckLocalDatabase
}
\examples{
"non"
}
