% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckMeta.R
\name{CheckMeta}
\alias{CheckMeta}
\title{CheckMeta : Checking samples accors MetaData files}
\usage{
CheckMeta(MetaData)
}
\arguments{
\item{MetaData}{a mMetaData list}
}
\value{
none
}
\description{
CheckMeta : Checking samples accors MetaData files
}
\examples{
"none"
}
