% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningClinic.R
\name{CleaningClinic}
\alias{CleaningClinic}
\title{CleaningClinic function : Cleaning clinical data to map samples and patients in Sequençing data}
\usage{
CleaningClinic(
  Metadata = NULL,
  ClinicToClean = NULL,
  Lexic = NULL,
  exportname = NULL,
  type = c("Samples", "Patients"),
  FilterSamples = F,
  FilterPatients = F,
  CleanFromOtherType = F,
  force.replace = F,
  keep.all.column = F
)
}
\arguments{
\item{Metadata}{Metadata object}

\item{ClinicToClean}{a character string of name of clinical data to clean.}

\item{Lexic}{Lexic to use for cleaning. Created from CreateLexic function .}

\item{exportname}{name to apply in Metadata object list}

\item{type}{"c("Samples", "Patients") for building clean clinical data from raw clinical data.}

\item{FilterSamples}{default F, if T, keep only retrieved samples in SamplesAnnotation file}

\item{FilterPatients}{default F, if T, keep only retrieved patients in Count SamplesAnnotation file}

\item{CleanFromOtherType}{If TRUE, Force a cleaning method from Samples.clinical data into Patient.clinical data and vice versa.}

\item{force.replace}{set as F. T : replace an already object with the same name}

\item{keep.all.column}{default F, if T, copy all column from clinic.}
}
\value{
a data frame of Samples pheno or patients clinical data. If Sample ID and Patients ID are the sames, so Samples.pheno and Patient_clinic are the same data frame
}
\description{
CleaningClinic function : Cleaning clinical data to map samples and patients in Sequençing data
}
\examples{
"none"
}
