% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningData.R
\name{CleaningData}
\alias{CleaningData}
\title{CleaningData from Metaobject}
\usage{
CleaningData(
  Metadata = NULL,
  PatientsLexic = NULL,
  PatientsAnnotToClean = NULL,
  PatientsExportname = NULL,
  SamplesLexic = NULL,
  SamplesAnnotToClean = NULL,
  SamplesExportname = NULL,
  FilterSP = F,
  force.replace = F,
  keep.all.column = F
)
}
\arguments{
\item{Metadata}{a Metadata Object to fill}

\item{PatientsLexic}{Lexic to use for cleaning. Created from CreateLexic function .}

\item{PatientsAnnotToClean}{a character string of name of clinical data to clean.}

\item{PatientsExportname}{name to apply in Metadata object list}

\item{SamplesLexic}{Lexic to use for cleaning. Created from CreateLexic function .}

\item{SamplesAnnotToClean}{a character string of name of Samples annotation data to clean.}

\item{SamplesExportname}{name to apply in Metadata object list}

\item{FilterSP}{default F, if T, keep only retrieved samples in SamplesAnnotation file}

\item{force.replace}{set as F. T : replace an already object with the same name}

\item{keep.all.column}{default F, if T, copy all column from clinic.}
}
\value{
a meataobject
}
\description{
CleaningData from Metaobject
}
\examples{
"non"
}
