% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColNameClinic.R
\name{ColNameClinic}
\alias{ColNameClinic}
\title{ColNameClinic function to prepare constant colnames to find in clinical data from lexiue .txt}
\usage{
ColNameClinic(Lexical_colnames_path)
}
\arguments{
\item{Lexical_colnames_path}{file path to find lexique of colnames}
}
\value{
a list
}
\description{
ColNameClinic function to prepare constant colnames to find in clinical data from lexiue .txt
}
\examples{
"none"
}
