% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeLexic.R
\name{completeLexic}
\alias{completeLexic}
\title{completeLexic}
\usage{
completeLexic(liste_de_listes, noms_listes)
}
\arguments{
\item{liste_de_listes}{Lexic}

\item{noms_listes}{colnames(clinic)}
}
\value{
un lexic
}
\description{
completeLexic
}
\examples{
none

}
