% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDataset.R
\name{CreateDataset}
\alias{CreateDataset}
\title{CreateDataset Create project Dir in Processes,n Precessed, Raw Clinic, Raw genomic, and references if does not exist.}
\usage{
CreateDataset(project, path)
}
\arguments{
\item{project}{a character string : project name}

\item{path}{dir path of Parent directory}
}
\value{
a file path list of dir project in ~parent/dir/project file
}
\description{
CreateDataset Create project Dir in Processes,n Precessed, Raw Clinic, Raw genomic, and references if does not exist.
}
\examples{
"none"



}
