% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLexic.R
\name{CreateLexic}
\alias{CreateLexic}
\title{CreateLexic function to prepare constant colnames to find in clinical data from lexiue .txt}
\usage{
CreateLexic(Dataset = NULL, type = c("SamplesLexic", "PatientsLexic"))
}
\arguments{
\item{Dataset}{file path to find lexique of colnames}

\item{type}{default F. If MetaObject is made of Single.Cell sequencing data, set to T.}
}
\value{
a list
}
\description{
CreateLexic function to prepare constant colnames to find in clinical data from lexiue .txt
}
\examples{
"none"
}
