% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportTSV.R
\name{ExportTSV}
\alias{ExportTSV}
\title{ExportTSV Export MetaData inside object into ".tsv" files}
\usage{
ExportTSV(MetaData, list.files.path, project)
}
\arguments{
\item{MetaData}{a MetaData  data files}

\item{list.files.path}{dirpath}

\item{project}{project}
}
\value{
".tsv" files into working directory
}
\description{
ExportTSV Export MetaData inside object into ".tsv" files
}
\examples{
"non"
}
