% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_attributes.R
\name{extract_attributes}
\alias{extract_attributes}
\title{Extract attributes from a gtf files to extract for example gene names ou type}
\usage{
extract_attributes(gtf_attributes, att_of_interest)
}
\arguments{
\item{gtf_attributes}{column attribute in the GTF file}

\item{att_of_interest}{a character specifying the targeted atribute to extract}
}
\value{
attribute of interest from the targeted line of data.frame with attributes
}
\description{
Extract attributes from a gtf files to extract for example gene names ou type
}
\examples{
"non"
}
