% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotation.R
\name{geneAnnotation}
\alias{geneAnnotation}
\title{geneAnnotation function to form a genes annotated data frame for data claening rnaseq data.}
\usage{
geneAnnotation(
  gtf.files = "gencode.v40.annotation.gtf",
  saverds = TRUE,
  file = ".rds"
)
}
\arguments{
\item{gtf.files}{GTF file to load, a character if wd is already set, or an object of GTF file already load}

\item{saverds}{if TRUE save a ".rds" file with the file field}

\item{file}{character to names the geneAnnotation rds file}

\item{type}{subset of genne Annot}

\item{GeneType}{target}

\item{GeneSymbol}{target}

\item{extract_attributes}{none}
}
\value{
return a data.frame with ensembl Id as rownames and attributes for each genes.
}
\description{
geneAnnotation function to form a genes annotated data frame for data claening rnaseq data.
}
\examples{
"none"

}
