% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProteinCodingMatrix.R
\name{getProteinCodingMatrix}
\alias{getProteinCodingMatrix}
\title{getProteinCodingMatrix}
\usage{
getProteinCodingMatrix(Metadata)
}
\arguments{
\item{Metadata}{Meta data from PublicDataNorm}
}
\value{
a Meta object
}
\description{
getProteinCodingMatrix
}
\examples{
"none"
}
