% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexicData.R
\name{LexicData}
\alias{LexicData}
\title{LexicData function to prepare constant colnames to find in clinical data from lexiue .txt}
\usage{
LexicData(list.files.path, MetaData = NULL, replaceLexic = F)
}
\arguments{
\item{list.files.path}{file path to find lexique of colnames}

\item{MetaData}{a MetaData list. Default NULL. Not mandatory. If Single.cell data, please specify your MetaData object to load specific lexic.}

\item{replaceLexic}{replace Samples and Patient Lexis if already loaded}
}
\value{
a list
}
\description{
LexicData function to prepare constant colnames to find in clinical data from lexiue .txt
}
\examples{
"none"
}
