% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathfiles.R
\name{Processpath}
\alias{Processpath}
\title{Processpath Export Dataset inside object into ".csv" files}
\usage{
Processpath(Dataset)
}
\arguments{
\item{Dataset}{a Dataset  data files}
}
\value{
a path to 02process project file
}
\description{
Processpath Export Dataset inside object into ".csv" files
}
\examples{
"non"
}
