% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PublicDataNorm-package.R
\docType{package}
\name{PublicDataNorm-package}
\alias{PublicDataNorm}
\alias{PublicDataNorm-package}
\title{PublicDataNorm: Building and cleaning of public RNAseq data}
\description{
Take public data : RNAseq raw and /or normalized, patient clinical data, gene annotation adn build a metafile of files.
}
\author{
\strong{Maintainer}: Jerome Raffenne \email{raffenne.jerome@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
