% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathfiles.R
\name{Refpath}
\alias{Refpath}
\title{Refpath Export Metadata inside object into ".csv" files}
\usage{
Refpath(Metadata)
}
\arguments{
\item{Metadata}{a Metadata  data files}
}
\value{
a path to 03Referecence project file
}
\description{
Refpath Export Metadata inside object into ".csv" files
}
\examples{
"non"
}
