% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removefilesrecursively.R
\name{removefilesrecursively}
\alias{removefilesrecursively}
\title{removefilesrecursively remove donwloaded files from the TCGA}
\usage{
removefilesrecursively(files)
}
\arguments{
\item{files}{list files in dir to remove}

\item{remove.files.recursively}{function}
}
\value{
nothing
}
\description{
removefilesrecursively remove donwloaded files from the TCGA
}
\examples{
  "none"
}
