% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetMetaData.R
\name{SubsetMetaData}
\alias{SubsetMetaData}
\title{SubsetMetaData subsettinf Metadata with samples of interests.}
\usage{
SubsetMetaData(Metadata, Samples.selection = NULL)
}
\arguments{
\item{Metadata}{meta object}

\item{Samples.selection}{string of characters samples}
}
\value{
a meta object
}
\description{
SubsetMetaData subsettinf Metadata with samples of interests.
}
\examples{
"none"
}
