% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalMeta.R
\name{SurvivalMeta}
\alias{SurvivalMeta}
\title{SurvivalMeta Sruvial OS or PFS analysis from MetaData}
\usage{
SurvivalMeta(Metadata, target, Group.sep = c("median", "Q1 et 3"), type)
}
\arguments{
\item{Metadata}{a metaData object}

\item{target}{gene ID}

\item{Group.sep}{median ou quantil 0.25/0.75 separation}

\item{type}{OS of PFS}
}
\value{
a survival plot and gene expression plus grouping in Patient's clinical data sheet.
}
\description{
SurvivalMeta Sruvial OS or PFS analysis from MetaData
}
\examples{
"non"
}
