% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimportEpxression.R
\name{TCGAimportEpxression}
\alias{TCGAimportEpxression}
\title{TCGAimportEpxression allow to import and build the rnaseq data from the TCGA}
\usage{
TCGAimportEpxression(
  Meta = NULL,
  project = c("TCGA-PAAD"),
  data.category = "Transcriptome Profiling",
  data.type = "Gene Expression Quantification",
  experimental.strategy = "RNA-Seq",
  sample.type = "Primary Tumor",
  platform = NULL,
  file.type = NULL,
  legacy = F,
  data.norm = c("Raw", "TPM", "FKPM"),
  remove.files.prepared = F
)
}
\arguments{
\item{Meta}{a Meta object. If NULL, will create the Meta object. If exist, will add objects from queries.}

\item{project}{a character of TCGA project name}

\item{data.category}{a character of data categorie}

\item{data.type}{a character of data type}

\item{experimental.strategy}{a character of experimental.strategy}

\item{sample.type}{a character of sample.type}

\item{platform}{platform}

\item{file.type}{filetype}

\item{legacy}{legacy T or F}

\item{data.norm}{a character "Raw"  = raw count, "TPM" = tpm normalisation gene expression, "FKPM" = fkpm normalisation gene expression if it exist}

\item{remove.files.prepared}{FALSE or TRUE to remove data downloaded}

\item{query}{query object exported}

\item{bcr_patient_barcode}{grouped by}
}
\value{
a data frame with build rna expression raw or normalized
}
\description{
TCGAimportEpxression allow to import and build the rnaseq data from the TCGA
}
\examples{
"none"

}
