% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionCheck.R
\name{VersionCheck}
\alias{VersionCheck}
\title{VersionCheck : Asks DataBaseSumary.txt file for the oldest project's version.}
\usage{
VersionCheck(Metadata, path = NULL)
}
\arguments{
\item{Metadata}{the Metaobject of the project}

\item{path}{file path to DataBaseSumary.txt}
}
\value{
Return a Versionned metadata object
}
\description{
VersionCheck : Asks DataBaseSumary.txt file for the oldest project's version.
}
\examples{
"non"
}
