% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddClinic.R
\name{AddClinic}
\alias{AddClinic}
\title{AddClinic to Meta object}
\usage{
AddClinic(
  Metadata,
  path,
  merge = c(F, T),
  mergeBy,
  name,
  name.local.file = NULL
)
}
\arguments{
\item{Metadata}{Meta object}

\item{path}{dir path in which the GDC project is saved, or local files are saved}

\item{merge}{merge loaded data clinic with existing clinial data : full_join by rownames}

\item{mergeBy}{colname using for merging clinical data.}

\item{name}{if loca=True, names to apply in Metadata object slot}

\item{name.local.file}{name file of interest in path directory, could be multiple names. c("a.csv","b.csv")}
}
\value{
a data.frame in the Meta Object
}
\description{
AddClinic to Meta object
}
\examples{
"none"
}
