% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddClinic.R
\name{AddClinic}
\alias{AddClinic}
\title{AddClinic to Meta object}
\usage{
AddClinic(Metadata, path, merge = c(F, T), name.local.file)
}
\arguments{
\item{Metadata}{Meta object}

\item{path}{dir path in which the GDC project is saved, or local files are saved}

\item{merge}{merge loaded data clinic with existing clinial data : cbind by rownames}

\item{name.local.file}{if loca=True, names to apply in Meatadata object slot}
}
\value{
a data.frame in the Meta Object
}
\description{
AddClinic to Meta object
}
\examples{
"none"
}
