% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddClinicFromFile.R
\name{AddClinicFromFile}
\alias{AddClinicFromFile}
\title{AddClinicFromFile to Meta object}
\usage{
AddClinicFromFile(
  Metadata,
  Raw.file.path,
  name.local.file = NULL,
  name,
  type = c("Samples", "Patients"),
  force.replace = F,
  Export = F,
  mergeToClinic = F,
  mergeBy,
  join = c("left_join", "full_join")
)
}
\arguments{
\item{Metadata}{Meta object}

\item{Raw.file.path}{dir path in which the GDC project is saved, or local files are saved}

\item{name.local.file}{name file of interest in path directory, could be multiple names. c("a.csv","b.csv")}

\item{name}{if loca=True, names to apply in Metadata object slot}

\item{type}{c("Samples", "Patients") To specify if clinical data refers to samples or patients data. Importat if mutliple samples for one patient.}

\item{force.replace}{set as F. T : replace an already object with the same name}

\item{Export}{TRUE or FALSE. If data to be Exported, set T.}

\item{mergeToClinic}{merge loaded data clinic with existing clincial data : full_join by rownames. False if you load more than 1 file.}

\item{mergeBy}{colname using for merging clinical data.}

\item{join}{c("left_join", "full_join")}
}
\value{
a data.frame in the Meta Object
}
\description{
AddClinicFromFile to Meta object
}
\examples{
"none"
}
