% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddClinicFromObject.R
\name{AddClinicFromObject}
\alias{AddClinicFromObject}
\title{AddClinicFromObject}
\usage{
AddClinicFromObject(
  Metadata,
  object,
  setID.Column = NULL,
  name,
  type = c("Samples", "Patients", "Cells"),
  Export = T,
  force.replace = F
)
}
\arguments{
\item{Metadata}{a Metadata Object to fill}

\item{object}{an object to add}

\item{setID.Column}{a character string : column to fetch colnames from Count matrix.}

\item{name}{name to apply in Metadata object list}

\item{type}{can be Samples", "Patients" or "Cells", to defiens Data.type attributes}

\item{Export}{TRUE or FALSE. If data to be Exported, set T.}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
a Metadata Object
}
\description{
AddClinicFromObject
}
\examples{
"none"

}
