% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddClinicFromObject.R
\name{AddClinicFromObject}
\alias{AddClinicFromObject}
\title{AddClinicFromObject}
\usage{
AddClinicFromObject(
  Metadata,
  object,
  name,
  SamplesFilter = F,
  type = c("Samples", "Patients", "Cells"),
  Export = c("Yes", "No"),
  force.replace = F
)
}
\arguments{
\item{Metadata}{a Metadata Object to fill}

\item{object}{an object to add}

\item{name}{name to apply in Metadata object list}

\item{SamplesFilter}{Default F, select only samples found in both the clinical data and the Expression Matrix.}

\item{type}{can be Samples", "Patients" or "Cells", to defiens Data.type attributes}

\item{Export}{data to export after cleaning c("Yes", "No")}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
a Metadata Object
}
\description{
AddClinicFromObject
}
\examples{
"none"

}
