% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddClinicFromObject.R
\name{AddClinicFromObject}
\alias{AddClinicFromObject}
\title{AddClinicFromObject}
\usage{
AddClinicFromObject(
  Metadata,
  object,
  name,
  SamplesFilter = F,
  Data.type = c("Clinic", "SamplesAnnot", "CellAnnot"),
  Export = c("Yes", "No")
)
}
\arguments{
\item{Metadata}{a Metadata Object to fill}

\item{object}{an object to add}

\item{name}{name to apply in Metadata objet list}

\item{SamplesFilter}{Default F, select only samples found in both the clinical data and the Expression Matrix.}

\item{Data.type}{attribute  c("Clinic", "SamplesAnnot","CellAnnot" )}

\item{Export}{data to export after cleaning c("Yes", "No")}
}
\value{
a Metadata Object
}
\description{
AddClinicFromObject
}
\examples{
"none"

}
