% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddExpressionMatrix.R
\name{AddExpressionMatrix}
\alias{AddExpressionMatrix}
\title{AddExpressionMatrix to Meta object}
\usage{
AddExpressionMatrix(
  Metadata,
  local = c(T, F),
  query,
  data.norm,
  path,
  name,
  Raw = T,
  name.local.file = NULL,
  force.replace = F
)
}
\arguments{
\item{Metadata}{Meta object}

\item{local}{T of F. If F, use query form. If F, add expression patrice from local file}

\item{query}{query file from TCGAimportExpression function}

\item{data.norm}{a character "Raw" ,"TPM", "FKPM"}

\item{path}{dir path in which the GDC project is saved, or local files are saved}

\item{name}{if loca=True, names to apply in Metadata object slot}

\item{Raw}{TRUE or FALSE. If Raw data, to be specified.}

\item{name.local.file}{name file of interest in path directory}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
a data.frame in the Meta Object
}
\description{
AddExpressionMatrix to Meta object
}
\examples{
"none"
}
