% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddExpressionMatrix.R
\name{AddExpressionMatrix}
\alias{AddExpressionMatrix}
\title{AddExpressionMatrix to Meta object}
\usage{
AddExpressionMatrix(
  Metadata,
  local = c(T, F),
  query,
  data.norm,
  path,
  name.local.file
)
}
\arguments{
\item{Metadata}{Meta object}

\item{local}{T of F. If F, use query form. If F, add expression patrice from local file}

\item{query}{query file from TCGAimportExpression function}

\item{data.norm}{a character "Raw" ,"TPM", "FKPM"}

\item{path}{dir path in which the GDC project is saved, or local files are saved}

\item{name.local.file}{if loca=True, names to apply in Meatadata object slot}
}
\value{
a data.frame in the Meta Object
}
\description{
AddExpressionMatrix to Meta object
}
\examples{
"none"
}
