% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddgeneAnnotFromObject.R
\name{AddgeneAnnotFromObject}
\alias{AddgeneAnnotFromObject}
\title{AddgeneAnnotFromObject to Metadata}
\usage{
AddgeneAnnotFromObject(Metadata, object, Filter.Genes = F, force.replace = F)
}
\arguments{
\item{Metadata}{Metadata object to add geneAnnotation}

\item{object}{object to add}

\item{Filter.Genes}{default F, if T, keep only retrieved genes in Count matrix}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
return a Metadata list of dataframe with geneAnntotation
}
\description{
AddgeneAnnotFromObject to Metadata
}
\examples{
"none"

}
