% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddgeneAnnotGtf.R
\name{AddgeneAnnotGtf}
\alias{AddgeneAnnotGtf}
\title{AddgeneAnnotGtf to Meta}
\usage{
AddgeneAnnotGtf(Meta, gtf.file.dir, gtf.files, force.replace = F)
}
\arguments{
\item{Meta}{Meta object to add geneAnnotation}

\item{gtf.file.dir}{set xorking directory to path to gtf.file}

\item{gtf.files}{gtf file to load}

\item{force.replace}{set as F. T : replace an already object with the same name}
}
\value{
return a Meta list of dataframe with geneAnntotation
}
\description{
AddgeneAnnotGtf to Meta
}
\examples{
"none"

}
