% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddLexic.R
\name{AddKeyLexic}
\alias{AddKeyLexic}
\title{AddLexic}
\usage{
AddKeyLexic(lexic, Param = c("IDname", "param target"))
}
\arguments{
\item{lexic}{PatientLexic or SamplesLexic}

\item{Param}{string of character c("A","B", etc) : "A" names of parameter, will be a colnames in cleaned clinic. "B"& etc targeted colname(s) from raw clinicil data}
}
\value{
a Lexic
}
\description{
AddLexic
}
\examples{
None
}
