% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddObjetToMeta.R
\name{AddObjetToMeta}
\alias{AddObjetToMeta}
\title{AddObjetToMeta}
\usage{
AddObjetToMeta(
  Metadata,
  object,
  Omics.type = c("RNAseq", "Single.Cell", "Microarray", "Spatial"),
  name,
  geneFilter = F,
  SamplesFilter = F,
  Data.type = c("Count", "Clinic", "SamplesAnnot", "geneAnnot", "CellAnnot"),
  Export = c("Yes", "No")
)
}
\arguments{
\item{Metadata}{a Metadata Object to fill}

\item{object}{an object to add}

\item{Omics.type}{one of these categories : "RNAseq", "Single.Cell", "Microarray", "Spatial".}

\item{name}{name to apply in Metadata objet list}

\item{geneFilter}{Default F, select only genes found in both the geneAnnot and the Expression Matrix.}

\item{SamplesFilter}{Default F, select only samples found in both the clinical data and the Expression Matrix.}

\item{Data.type}{attribute c("Count","Clinic", "SamplesAnnot","geneAnnot","CellAnnot" )}

\item{Export}{data to export after cleaning c("Yes", "No")}
}
\value{
a Metadata Object
}
\description{
AddObjetToMeta
}
\examples{
"none"
}
