% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckLocalDatabase.R
\name{CheckLocalDatabase}
\alias{CheckLocalDatabase}
\title{CheckLocalDatabase}
\usage{
CheckLocalDatabase(
  Local.Data.base.Path,
  Databasename = "DataBaseSummary.txt",
  Normalization.Method = NA,
  Sequencing.DeepenessPerMillionReads = NA,
  Sequencing.LengthPer.pb = NA,
  Sequencing.Run = NA,
  Sequencing.Platform = NA,
  First.Author = NA,
  DOI = NA,
  project = NA
)
}
\arguments{
\item{Local.Data.base.Path}{path to parent directory of the database.}

\item{Databasename}{name of the database.txt summary file}

\item{Normalization.Method}{Defaul = NA}

\item{Sequencing.DeepenessPerMillionReads}{Defaul = NA}

\item{Sequencing.LengthPer.pb}{Defaul = NA}

\item{Sequencing.Platform}{Defaul = NA}

\item{First.Author}{Defaul = NA}

\item{DOI}{doi of article if exist. Defaul = NA}

\item{project}{project name}
}
\value{
a .text tab delimited database summary
}
\description{
CheckLocalDatabase
}
\examples{
"non"
}
