% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckLocalDatabase.R
\name{CheckLocalDatabase}
\alias{CheckLocalDatabase}
\title{CheckLocalDatabase}
\usage{
CheckLocalDatabase(
  Local.Data.base.Path,
  Normalization.Method = NA,
  Technology = NA,
  Platform = NA,
  Run.spec = NA,
  First.Author = NA,
  DOI = NA,
  project = NA,
  Comment = NA,
  User = NA
)
}
\arguments{
\item{Local.Data.base.Path}{path to parent directory of the database.}

\item{Normalization.Method}{Defaul = NA}

\item{Technology}{Microarray, RNAseq, Single cell etc}

\item{Platform}{Brand of technology}

\item{Run.spec}{Deppenes, reads, etc etc}

\item{First.Author}{Defaul = NA}

\item{DOI}{doi of article if exist. Defaul = NA}

\item{project}{project name}

\item{Comment}{specify a comment for this export}

\item{User}{User who export the cleaned Meta object}

\item{Sequencing.DeepenessPerMillionReads}{Defaul = NA}
}
\value{
a .text tab delimited database summary
}
\description{
CheckLocalDatabase
}
\examples{
"non"
}
