% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckMeta.R
\name{CheckMeta}
\alias{CheckMeta}
\title{CheckMeta : Checking samples accors Metadata files}
\usage{
CheckMeta(Metadata)
}
\arguments{
\item{Metadata}{a Metadata list}
}
\value{
none
}
\description{
CheckMeta : Checking samples accors Metadata files
}
\examples{
"none"
}
