% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningClinic.R
\name{CleaningClinic}
\alias{CleaningClinic}
\title{CleaningClinic function : Cleaning clinical data to map samples and patients in Sequençing data}
\usage{
CleaningClinic(Metadata, type = c("Sample", "Patients"), Lexical_colnames_path)
}
\arguments{
\item{Metadata}{Metadata object}

\item{type}{"Sample Pheno" or "Patients' clinical data" for building clean clinical data from raw clinical data.}

\item{Lexical_colnames_path}{file path to find lexique of colnames}
}
\value{
a data frame of Samples pheno or patients clinical data. If Sample ID and Patients ID are the sames, so Samples.pheno and Patient_clinic are the same data frame
}
\description{
CleaningClinic function : Cleaning clinical data to map samples and patients in Sequençing data
}
\examples{
"none"
}
