% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleaningClinic.R
\name{CleaningClinic}
\alias{CleaningClinic}
\title{CleaningClinic function : Cleaning clinical data to map samples and patients in Sequençing data}
\usage{
CleaningClinic(
  Metadata,
  ClinicToClean = NULL,
  name = NULL,
  type = c("Samples", "Patients"),
  list.files.path,
  project,
  ForceCleaning = F,
  force.replace = F,
  all.col = F
)
}
\arguments{
\item{Metadata}{Metadata object}

\item{ClinicToClean}{a character string of name of clinical data to clean.}

\item{name}{name to apply in Metadata object list}

\item{type}{"c("Samples", "Patients") for building clean clinical data from raw clinical data.}

\item{list.files.path}{file path to find lexique of colnames}

\item{project}{project}

\item{ForceCleaning}{If TRUE, Force a cleaning method from Samples.clinical data into Patient.clinical data and vice versa.}

\item{force.replace}{set as F. T : replace an already object with the same name}

\item{all.col}{default F, if T, copy all column from clinic.}
}
\value{
a data frame of Samples pheno or patients clinical data. If Sample ID and Patients ID are the sames, so Samples.pheno and Patient_clinic are the same data frame
}
\description{
CleaningClinic function : Cleaning clinical data to map samples and patients in Sequençing data
}
\examples{
"none"
}
