% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirCheck.R
\name{DirCheck}
\alias{DirCheck}
\title{DirCheck Create project Dir in Script,n Precessed, Raw Clinic, Raw genomic, and references if does not exist.}
\usage{
DirCheck(project)
}
\arguments{
\item{project}{a character string : project name}
}
\value{
a file path list of dir project in ~parent/dir/project file
}
\description{
DirCheck Create project Dir in Script,n Precessed, Raw Clinic, Raw genomic, and references if does not exist.
}
\examples{
DirCheck("Maurer")



}
