% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCSV.R
\name{ExportCSV}
\alias{ExportCSV}
\title{ExportCSV Export Metadata inside object into ".csv" files}
\usage{
ExportCSV(Metadata)
}
\arguments{
\item{Metadata}{a Metadata  data files}
}
\value{
".csv" files into working directory
}
\description{
ExportCSV Export Metadata inside object into ".csv" files
}
\examples{
"non"
}
