% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCSV.R
\name{ExportCSV}
\alias{ExportCSV}
\title{ExportCSV Export MetaData inside object into ".csv" files}
\usage{
ExportCSV(MetaData, list.files.path, project)
}
\arguments{
\item{MetaData}{a Meta  data files}

\item{list.files.path}{dirpath}
}
\value{
".csv" files into working directory
}
\description{
ExportCSV Export MetaData inside object into ".csv" files
}
\examples{
"non"
}
