% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportTSV.R
\name{ExportTSV}
\alias{ExportTSV}
\title{ExportTSV Export Metadata inside object into ".tsv" files}
\usage{
ExportTSV(Metadata, encoding = "UTF-8")
}
\arguments{
\item{Metadata}{a Metadata  data files}

\item{encoding}{see file encoding section "UTF-8" or "latin1"}
}
\value{
".tsv" files into working directory
}
\description{
ExportTSV Export Metadata inside object into ".tsv" files
}
\examples{
"non"
}
