% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportTSV.R
\name{ExportTSV}
\alias{ExportTSV}
\title{ExportTSV Export Metadata inside object into ".tsv" files}
\usage{
ExportTSV(Metadata)
}
\arguments{
\item{Metadata}{a Metadata  data files}
}
\value{
".tsv" files into working directory
}
\description{
ExportTSV Export Metadata inside object into ".tsv" files
}
\examples{
"non"
}
