% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexicData.R
\name{LexicData}
\alias{LexicData}
\title{LexicData function to prepare constant colnames to find in clinical data from lexiue .txt}
\usage{
LexicData(list.files.path)
}
\arguments{
\item{list.files.path}{file path to find lexique of colnames}
}
\value{
a list
}
\description{
LexicData function to prepare constant colnames to find in clinical data from lexiue .txt
}
\examples{
"none"
}
