% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathfiles.R
\name{Processpath}
\alias{Processpath}
\title{Processpath Export Metadata inside object into ".csv" files}
\usage{
Processpath(Metadata)
}
\arguments{
\item{Metadata}{a Metadata  data files}
}
\value{
a path to 02process project file
}
\description{
Processpath Export Metadata inside object into ".csv" files
}
\examples{
"non"
}
