% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathfiles.R
\name{Rawpath}
\alias{Rawpath}
\title{Rawpath Export Dataset inside object into ".csv" files}
\usage{
Rawpath(Dataset)
}
\arguments{
\item{Dataset}{a Dataset  data files}
}
\value{
a path to 01RawData project file
}
\description{
Rawpath Export Dataset inside object into ".csv" files
}
\examples{
"non"
}
