% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathfiles.R
\name{Refpath}
\alias{Refpath}
\title{Refpath Export Dataset inside object into ".csv" files}
\usage{
Refpath(Dataset)
}
\arguments{
\item{Dataset}{a Dataset  data files}
}
\value{
a path to 03Referecence project file
}
\description{
Refpath Export Dataset inside object into ".csv" files
}
\examples{
"non"
}
