% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGAimportEpxression.R
\name{TCGAimportEpxression}
\alias{TCGAimportEpxression}
\title{TCGAimportEpxression allow to import and build the rnaseq data from the TCGA}
\usage{
TCGAimportEpxression(
  Meta = NULL,
  list.files.path = NULL,
  alt.dir = NULL,
  project = NULL,
  data.category = NULL,
  data.type = NULL,
  experimental.strategy = NULL,
  sample.type = NULL,
  platform = NULL,
  file.type = NULL,
  legacy = F,
  workflow.type = NULL,
  data.norm = c("Raw", "TPM", "FKPM", "FKPM_UQ")
)
}
\arguments{
\item{Meta}{a Meta object. If NULL, will create the Meta object. If exist, will add objects from queries.}

\item{list.files.path}{list.files.path}

\item{alt.dir}{aternaticve dir path for Windows user. Length of path can't exceed 250 character, Downaloading may be altered. The alt.dir path will be remove after moving GDCdata to list.file.path}

\item{project}{a character of TCGA project name}

\item{data.category}{a character of data categorie}

\item{data.type}{a character of data type}

\item{experimental.strategy}{a character of experimental.strategy}

\item{sample.type}{a character of sample.type}

\item{platform}{platform}

\item{file.type}{filetype}

\item{legacy}{legacy T or F}

\item{workflow.type}{workflow.type}

\item{data.norm}{a character "Raw"  = raw count, "TPM" = tpm normalisation gene expression, "FKPM" = fkpm normalisation gene expression if it exist}
}
\value{
a data frame with build rna expression raw or normalized
}
\description{
TCGAimportEpxression allow to import and build the rnaseq data from the TCGA
}
\examples{
"none"

}
