% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathfiles.R
\name{Verifiedpath}
\alias{Verifiedpath}
\title{Verifiedpath Export Metadata inside object into ".csv" files}
\usage{
Verifiedpath(Metadata)
}
\arguments{
\item{Metadata}{a Metadata  data files}
}
\value{
a path to 04VerifiedDataSet project file
}
\description{
Verifiedpath Export Metadata inside object into ".csv" files
}
\examples{
"non"
}
